AW fit interval with double values
git-svn-id: http://svn-ae/ae_doas/qdoas/trunk/Src@203 0551a244-8104-0410-bdc1-badfa85f4f5d

+++ mediator/mediate_analysis_window.h
@@ -44,8 +44,8 @@ extern "C" {
     char residualFile[FILENAME_BUFFER_LENGTH];
     double refSzaCenter;
     double refSzaDelta;
-    int fitMinWavelength;
-    int fitMaxWavelength;
+    double fitMinWavelength;
+    double fitMaxWavelength;
     int requireSpectrum;
     int requirePolynomial;
     int requireFit;
